set the previus homepage as the new root page
