+++ src/app/page.tsx
@@ -1,36 +1,23 @@
-import React from "react";
-import Button from "./components/button";
-import Image from "next/image";
-import Link from "next/link";
+"use client";
 
-const Home = () => {
+import SearchBar from "./components/search-bar";
+import HomepageTitle from "./components/homepage-title";
+import NavBar from "./components/nav-bar";
+import PopularSection from "./components/molecules/popular-section";
+import CategoriesSection from "./components/molecules/categories-section";
+
+const Homepage = () => {
+  const locationName = "Netherlands";
+
   return (
-    <div
-      style={{
-        backgroundImage: `url("/images/food-bowl.jpg")`,
-        backgroundSize: "cover",
-      }}
-      className="h-screen w-full px-8 py-11 flex flex-col justify-between"
-    >
-      <Image
-        src="/icons/chef-hat-icon.svg"
-        alt="chef hat icon"
-        height={164}
-        width={164}
-        className="invert self-center"
-      ></Image>
-
-      <div className="text-white font-montserrat flex flex-col gap-6">
-        <p className="text-2xl">
-          Find your <br />
-          <span className="text-5xl font-medium"> Next Meal</span>
-        </p>
-        <Link href={"homepage"}>
-          <Button text="Explore" />
-        </Link>
-      </div>
+    <div className="p-5 flex flex-col gap-6">
+      <NavBar />
+      <HomepageTitle userName="Frederico" />
+      <SearchBar placeholder="Find things to do" />
+      <PopularSection />
+      <CategoriesSection />
     </div>
   );
 };
 
-export default Home;
+export default Homepage;
